# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the BakingUp Backend API.",
        "title": "BakingUp Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/ingredient/addIngredientStock": {
            "post": {
                "description": "Add ingredient stock by using ingredient stock request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Add ingredient stock",
                "parameters": [
                    {
                        "description": "Ingredient Stock Request",
                        "name": "AddIngredientStockRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AddIngredientStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot add ingredient stock\"'",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/addIngredients": {
            "post": {
                "description": "Add ingredient by using ingredient request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Add ingredient",
                "parameters": [
                    {
                        "description": "Ingredient Request",
                        "name": "AddIngredientRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AddIngredientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot add ingredients",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/deleteIngredient": {
            "delete": {
                "description": "Delete an ingredient by using ingredient id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Delete an ingredient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient ID",
                        "name": "ingredient_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot delete an ingredient",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/deleteIngredientBatchNote": {
            "delete": {
                "description": "Delete ingredient batch note by ingredient note ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Delete ingredient batch note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient Note ID",
                        "name": "ingredient_note_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot delete ingredient batch note",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/deleteIngredientStock": {
            "delete": {
                "description": "Delete an ingredient stock by using ingredient stock id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Delete an ingredient stock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient Stock ID",
                        "name": "ingredient_stock_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    },
                    "400": {
                        "description": "Cannot delete an ingredient stock",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/getAddEditIngredientStockDetail": {
            "get": {
                "description": "Get add edit ingredient stock details by ingredient ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Get add edit ingredient stock details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient ID",
                        "name": "ingredient_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.AddEditIngredientStockDetail"
                        }
                    },
                    "400": {
                        "description": "Cannot get add edit ingredient stock detail",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/getAllIngredients": {
            "get": {
                "description": "Get all ingredients by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Get all ingredients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.IngredientList"
                        }
                    },
                    "400": {
                        "description": "Cannot get all ingredients",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/getIngredientDetail": {
            "get": {
                "description": "Get ingredient details by ingredient ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Get ingredient details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient ID",
                        "name": "ingredient_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.IngredientDetail"
                        }
                    },
                    "400": {
                        "description": "Cannot get ingredient detail",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
        "/ingredient/getIngredientStockDetail": {
            "get": {
                "description": "Get ingredient stock details by ingredient stock ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredient"
                ],
                "summary": "Get ingredient stock details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient stock ID",
                        "name": "ingredient_stock_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/domain.IngredientStockDetail"
                        }
                    },
                    "400": {
                        "description": "Cannot get ingredient stock detail",
                        "schema": {
                            "$ref": "#/definitions/http.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AddEditIngredientStockDetail": {
            "type": "object",
            "properties": {
                "ingredient_eng_name": {
                    "type": "string"
                },
                "ingredient_thai_name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "domain.AddIngredientRequest": {
            "type": "object",
            "properties": {
                "day_before_expire": {
                    "type": "string"
                },
                "img": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ingredient_eng_name": {
                    "type": "string"
                },
                "ingredient_thai_name": {
                    "type": "string"
                },
                "stock_less_than": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.AddIngredientStockRequest": {
            "type": "object",
            "properties": {
                "expiration_date": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "ingredient_brand": {
                    "type": "string"
                },
                "ingredient_id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "supplier": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.Ingredient": {
            "type": "object",
            "properties": {
                "expiration_status": {
                    "type": "string"
                },
                "ingredient_id": {
                    "type": "string"
                },
                "ingredient_name": {
                    "type": "string"
                },
                "ingredient_url": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "domain.IngredientDetail": {
            "type": "object",
            "properties": {
                "ingredient_less_than": {
                    "type": "integer"
                },
                "ingredient_name": {
                    "type": "string"
                },
                "ingredient_quantity": {
                    "type": "string"
                },
                "ingredient_url": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stock_amount": {
                    "type": "integer"
                },
                "stocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Stock"
                    }
                }
            }
        },
        "domain.IngredientList": {
            "type": "object",
            "properties": {
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Ingredient"
                    }
                }
            }
        },
        "domain.IngredientNote": {
            "type": "object",
            "properties": {
                "ingredient_note": {
                    "type": "string"
                },
                "ingredient_note_id": {
                    "type": "string"
                },
                "note_created_at": {
                    "type": "string"
                }
            }
        },
        "domain.IngredientStockDetail": {
            "type": "object",
            "properties": {
                "day_before_expire": {
                    "type": "string"
                },
                "ingredient_brand": {
                    "type": "string"
                },
                "ingredient_eng_name": {
                    "type": "string"
                },
                "ingredient_price": {
                    "type": "string"
                },
                "ingredient_quantity": {
                    "type": "string"
                },
                "ingredient_stock_url": {
                    "type": "string"
                },
                "ingredient_supplier": {
                    "type": "string"
                },
                "ingredient_thai_name": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.IngredientNote"
                    }
                }
            }
        },
...
        "domain.Stock": {
            "type": "object",
            "properties": {
                "expiration_date": {
                    "type": "string"
                },
                "expiration_status": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "stock_id": {
                    "type": "string"
                },
                "stock_url": {
                    "type": "string"
                }
            }
        },
...
        "http.response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}